Version 3.2.10 (June 22, 2021) - added tax rate to tax descriptions - show taxes for tax included on entry forms - fixed missing DCN when batch printing or emailing - added message to internal notes when batch emailing - fixed POS readonly access - fixed payments formatting issue when printing to queue - trap errors when translating files

+++ locale/sv/locales.pl
@@ -83,6 +83,8 @@ foreach $file (@progfiles) {
   
   if (-f "$file.missing") {
     eval { require "$file.missing"; };
+    
+    die if $@;
     unlink "$file.missing";
 
     for (keys %$missing) {
@@ -144,8 +146,8 @@ $self{subs} = {
     $english_sub = lc $key;
     
     $translated_sub = lc $text;
-    $english_sub =~ s/( |-|,|\/|\.$)/_/g;
-    $translated_sub =~ s/( |-|,|\/|\.$)/_/g;
+    $english_sub =~ s/( |-|,|\/|\.$|\\')/_/g;
+    $translated_sub =~ s/( |-|,|\/|\.$|\\')/_/g;
     print FH qq|  '$translated_sub'|.(' ' x (27-length($translated_sub))).qq| => '$english_sub',\n|;
   }
